BIKE

Category 1

NistyPQC.BIKE.Category1.generate_keys — Function

generate_keys([; seed])

Return a tuple (; ek, dk) consisting of an encapsulation key and the corresponding decapsulation key. The length of ek will be 1541 bytes and the length of dk 3114 bytes.

For a deterministic result, a seed of 32 bytes can be provided.

NistyPQC.BIKE.Category1.encapsulate_secret — Function

encapsulate_secret(ek[; m])

Return a tuple (; K, c) consisting of a shared secret K and a ciphertext c from which K can be recomputed with the decapsulation key dk that corresponds to ek. The parameter ek must be a valid encapsulation key of 1541 bytes. The length of K will be 32 bytes and the length of c 1573 bytes.

For a deterministic result, a plaintext m of 32 bytes can be provided.

NistyPQC.BIKE.Category1.decapsulate_secret — Function

decapsulate_secret(c, dk)

Return the secret key K in case of successful decapsulation. Otherwise implicitly reject, i.e. return a deterministic value K derived from c and dk. The parameter c must have 1573 bytes and dk 3114 bytes. Both for successful and failed decapsulation, K will have a length of 32 bytes.

Category 3

NistyPQC.BIKE.Category3.generate_keys — Function

generate_keys([; seed])

Return a tuple (; ek, dk) consisting of an encapsulation key and the corresponding decapsulation key. The length of ek will be 3083 bytes and the length of dk 6198 bytes.

For a deterministic result, a seed of 32 bytes can be provided.

NistyPQC.BIKE.Category3.encapsulate_secret — Function

encapsulate_secret(ek[; m])

Return a tuple (; K, c) consisting of a shared secret K and a ciphertext c from which K can be recomputed with the decapsulation key dk that corresponds to ek. The parameter ek must be a valid encapsulation key of 3083 bytes. The length of K will be 32 bytes and the length of c 3115 bytes.

For a deterministic result, a plaintext m of 32 bytes can be provided.

NistyPQC.BIKE.Category3.decapsulate_secret — Function

decapsulate_secret(c, dk)

Return the secret key K in case of successful decapsulation. Otherwise implicitly reject, i.e. return a deterministic value K derived from c and dk. The parameter c must have 3115 bytes and dk 6198 bytes. Both for successful and failed decapsulation, K will have a length of 32 bytes.

Category 5

NistyPQC.BIKE.Category5.generate_keys — Function

generate_keys([; seed])

Return a tuple (; ek, dk) consisting of an encapsulation key and the corresponding decapsulation key. The length of ek will be 5122 bytes and the length of dk 10276 bytes.

For a deterministic result, a seed of 32 bytes can be provided.

NistyPQC.BIKE.Category5.encapsulate_secret — Function

encapsulate_secret(ek[; m])

Return a tuple (; K, c) consisting of a shared secret K and a ciphertext c from which K can be recomputed with the decapsulation key dk that corresponds to ek. The parameter ek must be a valid encapsulation key of 5122 bytes. The length of K will be 32 bytes and the length of c 5154 bytes.

For a deterministic result, a plaintext m of 32 bytes can be provided.

NistyPQC.BIKE.Category5.decapsulate_secret — Function

decapsulate_secret(c, dk)

Return the secret key K in case of successful decapsulation. Otherwise implicitly reject, i.e. return a deterministic value K derived from c and dk. The parameter c must have 5154 bytes and dk 10276 bytes. Both for successful and failed decapsulation, K will have a length of 32 bytes.