Falcon

Category 1

NistyPQC.Falcon.Category1.generate_keys — Function

generate_keys()

Return a tuple (; sk, pk) consisting of a secret key and the corresponding public key. The length of sk will be 1281 bytes and the length of pk 897 bytes.

NistyPQC.Falcon.Category1.sign_message — Function

sign_message(msg, sk[; salt])

Return a signature sig computed from the message msg based on the secret key sk. The message msg may consist of arbitrarily many bytes, whereas sk must be a valid secret key of 1281 bytes. The length of sig will be 666 bytes.

Optionally, a salt of 40 bytes used for hashing can be provided.

NistyPQC.Falcon.Category1.verify_signature — Function

verify_signature(msg, sig, pk)

Check whether sig is a valid signature for msg under the public key pk. The message msg and potential signature sig may consist of arbitrarily many bytes, whereas pk must be a valid public key of 897 bytes.

Category 5

NistyPQC.Falcon.Category5.generate_keys — Function

generate_keys()

Return a tuple (; sk, pk) consisting of a secret key and the corresponding public key. The length of sk will be 2305 bytes and the length of pk 1793 bytes.

NistyPQC.Falcon.Category5.sign_message — Function

sign_message(msg, sk[; salt])

Return a signature sig computed from the message msg based on the secret key sk. The message msg may consist of arbitrarily many bytes, whereas sk must be a valid secret key of 2305 bytes. The length of sig will be 1280 bytes.

Optionally, a salt of 40 bytes used for hashing can be provided.

NistyPQC.Falcon.Category5.verify_signature — Function

verify_signature(msg, sig, pk)

Check whether sig is a valid signature for msg under the public key pk. The message msg and potential signature sig may consist of arbitrarily many bytes, whereas pk must be a valid public key of 1793 bytes.